build(deps): align the Python dependency chain on its newest compatible versions
The Python declarations had drifted apart: the 3.12 CI lock ran JAX 0.9.2
while the 3.11 and 3.13 locks ran 0.10.0, the runtime and queuewaves locks
used protobuf 7 while every development lock used protobuf 6, numpy was
2.4.4 in five locks and 2.4.6 in three, build was 1.4.2 in the release
tools and 1.5.0 in the publish tools, and uv.lock differed from the 3.12 CI
lock in 49 packages. Resolving with --upgrade moved 121 of 244 development
pins, including the majors protobuf 7, rich 15, rpds-py 2026.6, backrefs 8,
readme-renderer 46, argon2-cffi-bindings 26 and types-protobuf 7, plus JAX
0.11.2, numpy 2.5.3, SciPy 1.18.1, pandas 3.0.6 and pytest 9.1.1.

Declarations changed so that the resolver reaches the current releases:

- protobuf >= 7.35.1 and grpcio/grpcio-tools >= 1.84.0: the regenerated gRPC
  stubs (protobuf 7.35.1 gencode, grpcio 1.84.0) check both at import, and the
  previous floors (protobuf 6.33.5, grpcio 1.60) no longer described a
  working install;
- upper bounds on websockets (<18), pymdown-extensions (<13), pygments (<3)
  and twine (<8) admit their current releases;
- filelock < 4 in the dev extra: the julia CI lane installs julia-lock.txt on
  top of the dev lock and juliapkg 0.1.27 requires filelock < 4, so without
  the bound the julia lock backtracked to juliacall 0.9.24 / juliapkg 0.1.15;
- the pinned tool inputs move to pip-audit 2.10.1, build 1.6.1 (release and
  publish tools now agree), cyclonedx-bom 7.4.0, twine 7.0.0, pre-commit 4.6.2;
- ruff 0.16.8 in pyproject and the pre-commit hook (pre-commit-hooks v6.0.0).
  Ruff 0.16's format hook also selects Markdown and would rewrite 80 docs
  files that CI never formats; the hook keeps types_or python/pyi/jupyter;
- ci-tools.txt moves to maturin 1.15.0 with all 14 published file hashes;
  docs-tools.txt, installed over the dev lock with --no-deps, now pins the
  same versions and hashes as the dev lock.

Python 3.11 keeps numpy 2.4.6, SciPy 1.17.1, JAX 0.10.2, contourpy 1.3.3 and
pypet 0.6.1 because the newer releases require Python >= 3.12. Remaining
differences to uv.lock are environmental: uv.lock also resolves the studio
extra, whose streamlit 1.64 caps websockets < 17, and uv keeps rtamt 0.4.10
for 3.13 while pip-compile honours its python_requires < 3.13.

Code follow-ups: numpy 2.5 stubs make nine casts redundant and exposed an
integer spike count summed into a float64-typed array (now summed as
float64); a types-PyYAML cast is redundant; click 8.5 deprecates
CliRunner.isolated_filesystem, so one CLI test uses a private working
directory instead.

Verified locally: mypy --strict over src and tools (795 files), ruff 0.16.8
check and format, scoped pre-commit, gRPC stub --check, 94 lock-contract
tests, 2,760 + 2,211 focused tests over the upgraded surfaces (NN/JAX, gRPC,
queuewaves, STL, schemas, CLI, server, Studio, docs), strict MkDocs, a build
with build 1.6.1 passing twine 7.0.0 check --strict, pip-audit clean on eight
locks, capability manifest, SPDX, REUSE. The venv matches uv.lock exactly.
The full suite and the 3.11/3.13 legs run in CI.

Authored by Anulum Fortis & Arcane Sapience (protoscience@anulum.li)
Seat: ae6d

+++ src/scpn_phase_orchestrator/adapters/neurocore_bridge.py
@@ -276,7 +276,7 @@ class NeurocoreBridge:
             return np.zeros(self._n_layers)
 
         spikes_2d = self._spike_counts.reshape(self._n_layers, self._n_per)
-        layer_spikes: FloatArray = spikes_2d.sum(axis=1)
+        layer_spikes: FloatArray = spikes_2d.sum(axis=1, dtype=np.float64)
         return layer_spikes / (self._n_per * duration_s)
 
     def _step_rust(self, layer_currents: FloatArray, n_substeps: int) -> FloatArray:
